mchat_project/requirements.txt
Django>=4.0,<5.0
daphne>=4.0
python-telegram-bot>=20.0
asgiref>=3.0
httpx>=0.20
anyio>=3.0
pytz